+++ CHANGELOG.md
@@ -14,6 +14,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update gdk to 0.71.0
 - Update translations
 
+## Unreleased
+
+### Added
+- Lightning notifications
+- Empty lightning account
+- Exporting lightning logs
+
 ## [4.0.26] - 2024-03-19
 
 ### Changed